+++ functions/src/firebaseAdmin.ts
@@ -1,0 +1,7 @@
+import * as admin from 'firebase-admin'
+
+if (admin.apps.length === 0) {
+  admin.initializeApp()
+}
+
+export { admin }
